Split file for low mem like samd21, and slightly improve LUT code
